backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
opencv-python
pillow
pydantic
bcrypt
python-multipart 